StoreDataSource пилотная версия

+++ storeDataSource.ts
@@ -1,23 +1,19 @@
 import cloneDeep from "lodash.clonedeep";
 import {DataWithUuid} from "./type/dataWithUuid";
-import { v4 as uuidv4 } from 'uuid';
+import {v4 as uuidv4} from 'uuid';
 
 type ChangeDataSource<TItem extends DataWithUuid> = {
-    readonly changeType: 'addNewItem' | 'editItem' | 'deleteItem';
+    readonly changeType: 'addNewItem' | 'editItem' | 'deleteItem' | 'clear' | 'destroy';
     readonly itemsList: TItem[];
 }
 
 type CallbackApplyFilters<TItem extends DataWithUuid> = (dataList: TItem[]) => TItem[];
 type CallbackChangeDataSource<TItem extends DataWithUuid> = (param: ChangeDataSource<TItem>) => void;
 
-type InitData<TItem extends DataWithUuid> = {
-    readonly filters?: CallbackApplyFilters<TItem>;
-}
-
 /**
  * Хранилище для управления данными
  */
-export class StoreDataSource<TItem extends DataWithUuid> {
+export default class StoreDataSource<TItem extends DataWithUuid> {
     /**
      * Элементы для внутреннего хранения в формат Hash Tables
      * @private
@@ -29,6 +25,11 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @private
      */
     private _callbacksListenersChangeDataSource: Record<string, CallbackChangeDataSource<TItem>>;
+
+    /**
+     * Функция для фильтрации элементов
+     * @private
+     */
     private _callbackApplyFilters?: CallbackApplyFilters<TItem>;
 
     /**
@@ -38,8 +39,8 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @param item
      * @private
      */
-    private _validItem(item: TItem): TItem {
-        const copyItem:TItem = cloneDeep(item);
+    private _validNewItem(item: TItem): TItem {
+        const copyItem: TItem = cloneDeep(item);
 
         if (typeof copyItem.uuid !== 'string') {
             return {
@@ -66,14 +67,37 @@ export class StoreDataSource<TItem extends DataWithUuid> {
     }
 
     /**
+     * Проверить существующий элемент
+     * Если uuid не строка или пустая строка вернет undefined
+     * Если элемента нет в списке элементов вернет undefined
+     * @param item
+     * @private
+     */
+    private _validExistingItem(item: TItem): TItem | undefined {
+        if (typeof item.uuid !== 'string') {
+            return undefined;
+        }
+
+        if (!item.uuid) {
+            return undefined;
+        }
+
+        if (!this._internalItems.has(item.uuid)) {
+            return undefined;
+        }
+
+        return cloneDeep(item);
+    }
+
+    /**
      * Добавляет новый элемент к общему списку
      * @param newItem
      * @private
      */
-    private _addNewItem(newItem: TItem):TItem {
-        const validNewItem = this._validItem(newItem);
+    private _addNewItem(newItem: TItem): TItem {
+        const validNewItem = this._validNewItem(newItem);
         this._internalItems.set(validNewItem.uuid, validNewItem);
-        return validNewItem;
+        return cloneDeep(validNewItem);
     }
 
     /**
@@ -82,11 +106,11 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @private
      */
     private _ediItem(existingItem: TItem): TItem | undefined {
-        const validExistingItem = this._validItem(existingItem);
+        const validExistingItem = this._validExistingItem(existingItem);
 
-        if(this._internalItems.has(validExistingItem.uuid)){
+        if (validExistingItem) {
             this._internalItems.set(validExistingItem.uuid, validExistingItem);
-            return validExistingItem;
+            return cloneDeep(validExistingItem);
         }
 
         return undefined;
@@ -115,15 +139,15 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @private
      */
     private _getItemByUuid(uuid: string): undefined | TItem {
-        if(typeof uuid !== 'string'){
+        if (typeof uuid !== 'string') {
             return undefined;
         }
 
-        if(!uuid){
+        if (!uuid) {
             return undefined;
         }
 
-        if(this._internalItems.has(uuid)) {
+        if (this._internalItems.has(uuid)) {
             return cloneDeep(this._internalItems.get(uuid));
         }
 
@@ -154,6 +178,10 @@ export class StoreDataSource<TItem extends DataWithUuid> {
         }
     }
 
+    /**
+     * Применить функцию фильтрации если есть
+     * @private
+     */
     private _applyCallbackFilter(): TItem[] {
         const itemsArray: TItem[] = this._getItemsArray();
 
@@ -164,27 +192,32 @@ export class StoreDataSource<TItem extends DataWithUuid> {
         return itemsArray;
     }
 
-
-    constructor(initData?: InitData<TItem>) {
+    constructor() {
         this._internalItems = new Map<string, TItem>();
         this._callbacksListenersChangeDataSource = {};
-        this._callbackApplyFilters = initData?.filters;
+        this._callbackApplyFilters = undefined;
     }
 
     /**
      * Получить список элементов
+     * Если есть функция фильтрации будет применен фильтр
      */
     get itemsList(): TItem[] {
-        const arr: TItem[] = Array.from(this._internalItems.values());
-        return cloneDeep(arr);
+        return this._applyCallbackFilter();
     }
 
-    //region Добавить/удалить слушатель изменения данных
+    /**
+     * Количество элементов
+     */
+    get itemsCount(): number {
+        return this._internalItems.size;
+    }
+
     /**
      * Добавить слушатель изменения данных
      * @param listener
      */
-    public addChangeDataSourceListener(listener: CallbackChangeDataSource<TItem>){
+    public addListenerChangeDataSource(listener: CallbackChangeDataSource<TItem>) {
         const listenerId: string = uuidv4();
         this._callbacksListenersChangeDataSource[listenerId] = listener;
     }
@@ -193,7 +226,7 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * Удалить слушатель изменения данных
      * @param listener
      */
-    public removeChangeDataSourceListener(listener: CallbackChangeDataSource<TItem>) {
+    public removeListenerChangeDataSource(listener: CallbackChangeDataSource<TItem>) {
         for (const listenerId in this._callbacksListenersChangeDataSource) {
             if (listener === this._callbacksListenersChangeDataSource[listenerId]) {
                 delete this._callbacksListenersChangeDataSource[listenerId];
@@ -201,8 +234,21 @@ export class StoreDataSource<TItem extends DataWithUuid> {
             }
         }
     }
-    //endregion
 
+    /**
+     * Добавить функцию фильтрации
+     * @param callback - функция фильтрации
+     */
+    public addFilter(callback: CallbackApplyFilters<TItem>) {
+        this._callbackApplyFilters = callback;
+    }
+
+    /**
+     * Удалить функцию фильтрации
+     */
+    public removeFilter() {
+        this._callbackApplyFilters = undefined;
+    }
 
     /**
      * Получить элемент по его uuid
@@ -218,22 +264,79 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @param uuidList
      */
     public getItemsListByUuid(uuidList: string[]): TItem[] {
-        const result: TItem[] = [];
-
-        if(!Array.isArray(uuidList)){
-            return result;
+        if (!Array.isArray(uuidList)) {
+            return [];
         }
 
-        for(const uuid of uuidList){
+        const result: TItem[] = [];
+        for (const uuid of uuidList) {
             const item = this._getItemByUuid(uuid);
-            if(item){
+            if (item) {
                 result.push(item);
             }
         }
 
-        return result;
+        return cloneDeep(result);
     }
 
+    /**
+     * Установить новый источник данных
+     * При установке очистит старый набор данных
+     * @param itemsList - Список элементов
+     * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
+     */
+    public setNewDataSource(itemsList: TItem[], isWithoutTrigger?: boolean) {
+        // Очищаем текущий источник данных
+        this._internalItems.clear();
+        // Добавляем элементы
+        for (const item of itemsList) {
+            const validItem = this._validNewItem(item);
+            this._internalItems.set(validItem.uuid, validItem);
+        }
+
+        if (!isWithoutTrigger) {
+            this._applyCallbacksListenersChangeDataSource({
+                changeType: 'addNewItem',
+                itemsList: this._applyCallbackFilter()
+            });
+        }
+    }
+
+    /**
+     * Очищает текущий набор данных
+     * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
+     */
+    public clearDataSource(isWithoutTrigger?: boolean) {
+        this._internalItems.clear();
+
+        if (!isWithoutTrigger) {
+            this._applyCallbacksListenersChangeDataSource({
+                changeType: 'clear',
+                itemsList: this._applyCallbackFilter()
+            });
+        }
+    }
+
+    /**
+     * Очищает все переданные данные
+     * Будет очищен набор хранимых элементов, все слушатели, функция фильтрации
+     * @param isWithoutTrigger
+     */
+    public destroy(isWithoutTrigger?: boolean) {
+        // Забываем все хранимые элементы
+        this._internalItems.clear();
+        // Очищаем всех слушателей
+        this._callbacksListenersChangeDataSource = {};
+        // Забываем функцию фильтрации
+        this._callbackApplyFilters = undefined;
+
+        if (!isWithoutTrigger) {
+            this._applyCallbacksListenersChangeDataSource({
+                changeType: 'destroy',
+                itemsList: this._applyCallbackFilter()
+            });
+        }
+    }
 
     /**
      * Удалить существующий элемент по uuid
@@ -242,14 +345,16 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @param uuid - uuid элемента который нужно удалить
      * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
      */
-    public deleteItemByUuid(uuid: string, isWithoutTrigger?: boolean ): boolean {
+    public deleteItemByUuid(uuid: string, isWithoutTrigger?: boolean): boolean {
         const isDeleteItem: boolean = this._deleteItemByUuid(uuid);
 
-        if(!isWithoutTrigger) {
-            this._applyCallbacksListenersChangeDataSource({
-                changeType: 'deleteItem',
-                itemsList: this._applyCallbackFilter()
-            });
+        if (!isWithoutTrigger) {
+            if (isDeleteItem) {
+                this._applyCallbacksListenersChangeDataSource({
+                    changeType: 'deleteItem',
+                    itemsList: this._applyCallbackFilter()
+                });
+            }
         }
 
         return isDeleteItem
@@ -267,18 +372,19 @@ export class StoreDataSource<TItem extends DataWithUuid> {
             return false;
         }
 
-        let isDeleteAllSuccess: boolean = true;
+        const needDeleteItemsCount: number = uuidList.length;
+        let successDeleteItemsCount: number = 0;
 
         for (const uuid of uuidList) {
             const isDeleteItemSuccess: boolean = this._deleteItemByUuid(uuid);
 
-            if (!isDeleteItemSuccess) {
-                isDeleteAllSuccess = false;
+            if (isDeleteItemSuccess) {
+                successDeleteItemsCount++;
             }
         }
 
         if (!isWithoutTrigger) {
-            if (isDeleteAllSuccess) {
+            if (successDeleteItemsCount > 0) {
                 this._applyCallbacksListenersChangeDataSource({
                     changeType: 'deleteItem',
                     itemsList: this._applyCallbackFilter()
@@ -286,7 +392,7 @@ export class StoreDataSource<TItem extends DataWithUuid> {
             }
         }
 
-        return isDeleteAllSuccess;
+        return needDeleteItemsCount === successDeleteItemsCount;
     }
 
     /**
@@ -294,10 +400,10 @@ export class StoreDataSource<TItem extends DataWithUuid> {
      * @param item - новый элемент который нужно добавить
      * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
      */
-    public addNewItem(item: TItem, isWithoutTrigger?: boolean ): TItem {
-       const newItem = this._addNewItem(item);
+    public addNewItem(item: TItem, isWithoutTrigger?: boolean): TItem {
+        const newItem = this._addNewItem(item);
 
-        if(!isWithoutTrigger) {
+        if (!isWithoutTrigger) {
             this._applyCallbacksListenersChangeDataSource({
                 changeType: 'addNewItem',
                 itemsList: this._applyCallbackFilter()
@@ -310,43 +416,44 @@ export class StoreDataSource<TItem extends DataWithUuid> {
     /**
      * Добавить список новых элементов
      * @param itemsList - список новых элементов которые необходимо добавить
-     * @param isWithoutTrigger
+     * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
      */
-    public addNewItemsList(itemsList: TItem[], isWithoutTrigger?: boolean ): TItem[] {
-        const addedItemsList: TItem[] = [];
-
-        if(!Array.isArray(itemsList)) {
-            return addedItemsList;
+    public addNewItemsList(itemsList: TItem[], isWithoutTrigger?: boolean): TItem[] {
+        if (!Array.isArray(itemsList)) {
+            return [];
         }
 
-        for(const item of itemsList){
+        const addedItemsList: TItem[] = [];
+
+        for (const item of itemsList) {
             const addedItem = this._addNewItem(item);
             addedItemsList.push(addedItem);
         }
 
-        if(!isWithoutTrigger) {
-            this._applyCallbacksListenersChangeDataSource({
-                changeType: 'addNewItem',
-                itemsList: this._applyCallbackFilter()
-            });
+        if (!isWithoutTrigger) {
+            if (addedItemsList.length) {
+                this._applyCallbacksListenersChangeDataSource({
+                    changeType: 'addNewItem',
+                    itemsList: this._applyCallbackFilter()
+                });
+            }
         }
 
         return addedItemsList;
     }
 
-
     /**
      * Редактировать существующий элемент
      * Вернет элемент если его удалось найти в общем списке
      * Если элемент не удалось найти вернет undefined
-     * @param item
-     * @param isWithoutTrigger
+     * @param item - Элемент который нужно отредактировать
+     * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
      */
-    public editItem(item: TItem, isWithoutTrigger?: boolean ): TItem | undefined {
+    public editItem(item: TItem, isWithoutTrigger?: boolean): TItem | undefined {
         const editItem: TItem | undefined = this._ediItem(item);
 
-        if(!isWithoutTrigger) {
-            if (editItem){
+        if (!isWithoutTrigger) {
+            if (editItem) {
                 this._applyCallbacksListenersChangeDataSource({
                     changeType: 'editItem',
                     itemsList: this._applyCallbackFilter()
@@ -357,133 +464,38 @@ export class StoreDataSource<TItem extends DataWithUuid> {
         return editItem;
     }
 
+    /**
+     * Редактировать список существующих элементов
+     * Вернет массив измененных элементов
+     * Если не удалось изменить не один элемент вернет пустой массив
+     * @param itemsList - список элементов которые нужно изменить
+     * @param isWithoutTrigger - триггеры слушателей не сработают если флаг будет установлен в true
+     */
+    public editItemsList(itemsList: TItem[], isWithoutTrigger?: boolean): TItem[] {
 
+        if (!Array.isArray(itemsList)) {
+            return [];
+        }
 
+        const changedItemsList: TItem[] = [];
 
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    // /**
-    //  * Сбросить все данные хранилища
-    //  */
-    // public resetStoreData() {
-    //     this._dataForManage = [];
-    // }
-    //
-    //
-    //
-    //
-    // //region Данные для управления
-    //
-    // /**
-    //  * Получить все текущие управляемые данные
-    //  */
-    // get allDataForManage() {
-    //     return this._dataForManage.slice();
-    // }
-    //
-    // /**
-    //  * Получить текущие управляемые данные
-    //  */
-    // get dataForManage() {
-    //     const copyData = this._dataForManage.slice();
-    //
-    //     if (typeof this._applyFilters === 'function') {
-    //         // Если есть функция фильтрации, фильтруем данные и отдаем их
-    //         return this._applyFilters(copyData);
-    //     }
-    //
-    //     // Если функции фильтрации нет, просто отдаем данные
-    //     return copyData;
-    // }
-    //
-    // /**
-    //  * Установлены данные или нет
-    //  */
-    // get isDataSet(): boolean {
-    //     return !!this._dataForManage.length;
-    // }
-    //
-    // /**
-    //  * Установить данные для управления
-    //  * @param dataList
-    //  */
-    // set dataForManage(dataList: DataItem[]) {
-    //     this._dataForManage = dataList.slice();
-    // }
-    //
-    // //endregion
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    // //region Функция фильтрации данных
-    // private _applyFilters?: CallbackApplyFilters<DataItem>;
-    //
-    // /**
-    //  * Установить функцию фильтрации
-    //  * @param callbackApplyFilters
-    //  */
-    // public setDataFilters(callbackApplyFilters: CallbackApplyFilters<DataItem>) {
-    //     this._applyFilters = callbackApplyFilters
-    // }
-    //
-    // //endregion
-    //
-    //
-    //
-    //
-    //
-    //
-    //
-    // /**
-    //  * Получить uuid элемента по атрибуту data-uuid HTML элемента
-    //  * @param element
-    //  */
-    // public getUuidByDataAttributeHTMLElement(element: HTMLElement): string | null {
-    //     return element.getAttribute('data-id');
-    // }
-    //
-    // /**
-    //  * Найти элемент по атрибуту data-uuid
-    //  * @param element
-    //  */
-    // public getItemByDataAttributeUuid(element: HTMLElement): DataItem | undefined {
-    //     const uuid: string | null = this.getUuidByDataAttributeHTMLElement(element);
-    //
-    //     if (!uuid) {
-    //         return undefined;
-    //     }
-    //
-    //     const foundElement = this.getItemByUuid(uuid);
-    //
-    //     if (!foundElement) {
-    //         return undefined;
-    //     }
-    //
-    //     return foundElement;
-    // }
+        for (const item of itemsList) {
+            const changedItem = this._ediItem(item);
 
+            if (changedItem) {
+                changedItemsList.push(changedItem);
+            }
+        }
 
+        if (!isWithoutTrigger) {
+            if (changedItemsList.length) {
+                this._applyCallbacksListenersChangeDataSource({
+                    changeType: 'editItem',
+                    itemsList: this._applyCallbackFilter()
+                });
+            }
+        }
+
+        return changedItemsList;
+    }
 }
